More work on getting Math::Simple to test

+++ t/conf-fuzz.t
@@ -1,6 +1,7 @@
 use strict;
 use warnings;
 
+use IPC::Run3;
 use IPC::System::Simple qw(system);
 use Test::Most;
 use Test::Needs 'Math::Simple';
@@ -20,22 +21,7 @@ my $contents = do { local $/; <$fh> };
 close $fh;
 
 like($contents, qr/diag\(/, 'fuzz test has diag line');
-
-# Auto-detect all test names
-open $fh, '<', $outfile or die "$outfile: $!";
-my @lines = <$fh>;
-close $fh;
-chomp @lines;
-
-my @detected_tests;
-
-for my $line (@lines) {
-	if ($line =~ /\b(?:ok|is|like|unlike)\s*\(.*?,\s*['"](.+?)['"]\s*\)/) {
-		push @detected_tests, $1;
-	}
-}
-
-ok(@detected_tests, 'Detected at least one test in the generated file');
+like($contents, qr/add/, 'mentions function under test');
 
 eval {
 	system("$^X -c $outfile");
@@ -43,9 +29,26 @@ eval {
 
 if($@) {
 	diag($@);
+	fail("$outfile compiles");
 } else {
-	unlink $outfile;
+	pass("$outfile compiles");
+
+	# Run the generated test
+	my ($stdout, $stderr);
+	run3 [ $^X, $outfile ], undef, \$stdout, \$stderr;
+
+	ok($? == 0, 'Generated test script exits successfully');
+
+	if($? == 0) {
+		unlink $outfile;
+	} else {
+		diag("STDERR:\n$stderr");
+		diag("STDOUT:\n$stdout");
+	}
+
+	like($stderr, qr/Math::Simple->add test case created/);
+	like($stdout, qr/^ok \d/sm, 'At least one created test passed');
+	unlike($stdout, qr/^not ok \d/sm, 'No created test failed');
 }
-ok(!$@, "$outfile compiles");
 
 done_testing();
